+++ src/app.tsx
@@ -1,0 +1,77 @@
+import { useSnackbar } from 'components/Feedback';
+import { Text } from 'components/Common';
+import { initAmplify } from 'core/amplify';
+import { appConfig } from 'core/configs';
+import { parseUrlSearchStr } from 'core/navigation';
+import React, { FC, useEffect } from 'react';
+import { Screens } from 'screens';
+import { mx, Styles } from 'styles';
+import { Log, capitalizeFirstLetter } from 'utils';
+import { useAuth } from 'core/api';
+
+const log = Log('app');
+
+initAmplify();
+
+export const App: FC = () => {
+  const { loaded: authLoaded, user: authUser } = useAuth();
+  const { showSnackbar } = useSnackbar();
+
+  useEffect(() => {
+    processIncomingQsParams();
+  }, []);
+
+  // Incoming params
+
+  const processIncomingQsParams = () => {
+    const params = parseUrlSearchStr(location.search);
+    if (params.action === 'cognitoConfirmUser') {
+      processCognitoConfirmUserAction(params);
+    }
+    if (params.action === 'cognitoConfirmAttr') {
+      processCognitoConfirmAttrAction(params);
+    }
+  };
+
+  const processCognitoConfirmUserAction = (params: Record<string, string>) => {
+    log.info('processing incoming cognitoConfirmUser action, params=', params);
+    if (params.err) {
+      showSnackbar(params.err, 'error');
+    } else {
+      showSnackbar('User has been confirmed!', 'success');
+    }
+  };
+
+  const processCognitoConfirmAttrAction = (params: Record<string, string>) => {
+    log.info('processing incoming cognitoConfirmEmail action, params=', params);
+    const { err, name = 'email' } = params;
+    if (err) {
+      showSnackbar(`Confirming ${name} error: ${err}`, 'error');
+    } else {
+      showSnackbar(`${capitalizeFirstLetter(name)} has been confirmed!`, 'success');
+    }
+  };
+
+  // Render
+
+  return (
+    <>
+      {authLoaded && <Screens logined={!!authUser} />}
+      {appConfig.env !== 'prd' && (
+        <Text style={styles.envLabel} block={true}>{`v${appConfig.version} (${appConfig.env})`}</Text>
+      )}
+    </>
+  );
+};
+
+const styles: Styles = {
+  envLabel: {
+    position: 'fixed',
+    right: 8,
+    bottom: 8,
+    ...mx.font(10, 'inherit', 500),
+    zIndex: 100,
+  },
+};
+
+export default App;