+++ src/components/Dashboard/UserNav/index.tsx
@@ -1,7 +1,7 @@
 import { Grid, makeStyles, Theme, useTheme } from '@material-ui/core';
 import { Image } from 'components/Common';
 import React, { FC, MouseEvent } from 'react';
-import { StyleProps } from 'styles';
+import { colors, Style, StyleProps } from 'styles';
 
 import addIcon from './assets/addIcon.svg';
 import contactIcon from './assets/contactIcon.svg';
@@ -12,13 +12,19 @@ import DashboardButton from './components/Button';
 
 interface Props extends StyleProps {
   disabledBtns?: BtnType[];
+  btnsBackgroundColor?: string;
   hiddenBtns?: BtnType[];
   onBtnClick?: (type: BtnType) => void;
 }
 
 type BtnType = 'add' | 'profile' | 'faq' | 'support' | 'contact';
 
-export const DashboardUserNav: FC<Props> = ({ disabledBtns, hiddenBtns, onBtnClick }) => {
+export const DashboardUserNav: FC<Props> = ({
+  disabledBtns,
+  hiddenBtns,
+  btnsBackgroundColor = colors.paleGrey,
+  onBtnClick,
+}) => {
   const isDisabled = (btn: BtnType) => disabledBtns && disabledBtns.includes(btn);
   const isHidden = (btn: BtnType) => hiddenBtns && hiddenBtns.includes(btn);
 
@@ -32,15 +38,23 @@ export const DashboardUserNav: FC<Props> = ({ disabledBtns, hiddenBtns, onBtnCli
   const theme = useTheme();
   const classes = useStyles(theme);
 
+  const btnStye: Style = { backgroundColor: btnsBackgroundColor };
+
   return (
     <Grid container className={classes.container}>
       {!isHidden('add') && (
-        <DashboardButton icon={<Image source={addIcon} />} onClick={handleClick('add')} disabled={isDisabled('add')}>
+        <DashboardButton
+          style={btnStye}
+          icon={<Image source={addIcon} />}
+          onClick={handleClick('add')}
+          disabled={isDisabled('add')}
+        >
           {'Add Event'}
         </DashboardButton>
       )}
       {!isHidden('profile') && (
         <DashboardButton
+          style={btnStye}
           icon={<Image source={profileIcon} />}
           onClick={handleClick('profile')}
           disabled={isDisabled('profile')}
@@ -49,12 +63,18 @@ export const DashboardUserNav: FC<Props> = ({ disabledBtns, hiddenBtns, onBtnCli
         </DashboardButton>
       )}
       {!isHidden('faq') && (
-        <DashboardButton icon={<Image source={faqIcon} />} onClick={handleClick('faq')} disabled={isDisabled('faq')}>
+        <DashboardButton
+          style={btnStye}
+          icon={<Image source={faqIcon} />}
+          onClick={handleClick('faq')}
+          disabled={isDisabled('faq')}
+        >
           {'FAQ'}
         </DashboardButton>
       )}
       {!isHidden('support') && (
         <DashboardButton
+          style={btnStye}
           icon={<Image source={supportIcon} />}
           onClick={handleClick('support')}
           disabled={isDisabled('support')}
@@ -64,6 +84,7 @@ export const DashboardUserNav: FC<Props> = ({ disabledBtns, hiddenBtns, onBtnCli
       )}
       {!isHidden('contact') && (
         <DashboardButton
+          style={btnStye}
           icon={<Image source={contactIcon} />}
           onClick={handleClick('contact')}
           disabled={isDisabled('contact')}
@@ -88,4 +109,5 @@ export const useStyles = (theme: Theme) =>
   })();
 
 export type DashboardUserNavBtnType = BtnType;
+export type DashboardUserNavProps = Props;
 export default DashboardUserNav;