+++ src/core/api/auth/context.tsx
@@ -1,19 +1,27 @@
 import Auth, { CognitoUser } from '@aws-amplify/auth';
-import { appConfig } from 'core/configs';
 import React, { createContext, FC, useContext, useEffect, useState } from 'react';
-import { errToStr, Log } from 'utils';
+import { Log } from 'utils';
 
-import { authGetCurrentUser, AuthSignUpInput } from './utils';
+import {
+  getCognitoCurUser,
+  CognitoSignUpInput,
+  cogntitoUserSignUp,
+  getCognitoMetadata,
+  isCognitoUserEmailComfirmed,
+  verifyCognitoEmail,
+} from './utils';
 
 const log = Log('core.api.auth.context');
 
 interface AuthContext {
   loaded: boolean;
+  userConfirmed: boolean;
   user?: CognitoUser;
   signIn: (email: string, password: string) => Promise<CognitoUser>;
-  signUp: (data: AuthSignUpInput) => Promise<CognitoUser>;
+  signUp: (data: CognitoSignUpInput) => Promise<CognitoUser>;
   forgotPassowrd: (email: string) => Promise<void>;
   forgotPasswordSubmit: (username: string, code: string, pass: string) => Promise<void>;
+  resendEmailConfirmation: () => Promise<void>;
   signOut: () => void;
 }
 
@@ -23,37 +31,16 @@ const mockFn = () => {
   throw new Error('Mock function');
 };
 
-const getClientMetadata = () => ({
-  type: 'web',
-  version: appConfig.version,
-  env: appConfig.env,
-  url: appConfig.url,
-});
-
-interface CognitoSignUpResponse {
-  user: CognitoUser;
-}
-
-const authUserSignUp = async ({ email, firstName, lastName, password }: AuthSignUpInput): Promise<CognitoUser> => {
-  log.info('signup with firstName=', firstName, ', lastName=', lastName, ', email=', email);
-  const attributes = { email, 'custom:firstName': firstName, 'custom:lastName': lastName };
-  const { user } = (await Auth.signUp({
-    username: email,
-    password,
-    attributes,
-    clientMetadata: getClientMetadata(),
-  })) as CognitoSignUpResponse;
-  return user;
-};
-
 export const useAuth = (): AuthContext => {
   const val = useContext(Context);
-  const mockData = {
+  const mockData: AuthContext = {
+    userConfirmed: false,
     loaded: false,
     signIn: mockFn,
     signUp: mockFn,
     forgotPassowrd: mockFn,
     forgotPasswordSubmit: mockFn,
+    resendEmailConfirmation: mockFn,
     signOut: () => undefined,
   };
   return val ? val : mockData;
@@ -62,6 +49,7 @@ export const useAuth = (): AuthContext => {
 export const AuthProvider: FC = ({ children }) => {
   const [loaded, setLoaded] = useState<boolean>(false);
   const [user, setUser] = useState<CognitoUser | undefined>(undefined);
+  const [userConfirmed, setUserConfirmed] = useState<boolean>(false);
 
   useEffect(() => {
     loadCurUser();
@@ -70,9 +58,13 @@ export const AuthProvider: FC = ({ children }) => {
   const loadCurUser = async () => {
     log.info('loading cur user');
     try {
-      const user = await getCurUser();
-      log.info(user ? 'user authrized' : 'user not authorized');
-      setUser(user);
+      const curUser = await getCognitoCurUser();
+      log.info(curUser ? 'user authrized' : 'user not authorized');
+      if (curUser) {
+        const isEmailConfirmed = await isCognitoUserEmailComfirmed(curUser);
+        setUserConfirmed(isEmailConfirmed);
+      }
+      setUser(curUser);
       setLoaded(true);
     } catch (err) {
       setLoaded(true);
@@ -80,36 +72,48 @@ export const AuthProvider: FC = ({ children }) => {
     }
   };
 
-  const getCurUser = async (): Promise<CognitoUser | undefined> => {
-    try {
-      return await Auth.currentAuthenticatedUser();
-    } catch (err) {
-      if (['The user is not authenticated', 'No current user'].includes(errToStr(err) || '')) {
-        return undefined;
-      } else {
-        throw err;
-      }
-    }
-  };
+  // Sign in
 
   const signIn = async (email: string, password: string): Promise<CognitoUser> => {
     const user = await Auth.signIn(email, password);
+    // Sending email verification after sign in
+    const isEmailConfirmed = await isCognitoUserEmailComfirmed(user);
+    if (!isEmailConfirmed) {
+      await verifyCognitoEmail(user);
+    }
+    setUserConfirmed(isEmailConfirmed);
     setUser(user);
     return user;
   };
 
-  const forgotPassowrd = async (email: string) => Auth.forgotPassword(email, getClientMetadata());
+  // Sign up
 
-  const forgotPasswordSubmit = async (username: string, code: string, pass: string) =>
-    Auth.forgotPasswordSubmit(username, code, pass, getClientMetadata());
-
-  const signUp = async (data: AuthSignUpInput): Promise<CognitoUser> => {
-    const user = await authUserSignUp(data);
+  const signUp = async (data: CognitoSignUpInput): Promise<CognitoUser> => {
+    const user = await cogntitoUserSignUp(data);
     return user;
   };
 
+  // Forgot password
+
+  const forgotPassowrd = async (email: string) => Auth.forgotPassword(email, getCognitoMetadata());
+
+  const forgotPasswordSubmit = async (username: string, code: string, pass: string) =>
+    Auth.forgotPasswordSubmit(username, code, pass, getCognitoMetadata());
+
+  // Email confirmation
+
+  const resendEmailConfirmation = async (): Promise<void> => {
+    if (!user) {
+      log.warn('trying to resend email confirmation while user is empty');
+      return;
+    }
+    return verifyCognitoEmail(user);
+  };
+
+  // Sign out
+
   const signOut = () => {
-    authGetCurrentUser()
+    getCognitoCurUser()
       .then(curUser => {
         if (curUser) {
           curUser.signOut();
@@ -125,7 +129,19 @@ export const AuthProvider: FC = ({ children }) => {
   };
 
   return (
-    <Context.Provider value={{ loaded, user, signIn, signUp, signOut, forgotPassowrd, forgotPasswordSubmit }}>
+    <Context.Provider
+      value={{
+        loaded,
+        userConfirmed,
+        user,
+        signIn,
+        signUp,
+        signOut,
+        forgotPassowrd,
+        forgotPasswordSubmit,
+        resendEmailConfirmation,
+      }}
+    >
       {children}
     </Context.Provider>
   );