+++ src/screens/Dashboard/Events/index.tsx
@@ -1,6 +1,8 @@
 import { Grid, Hidden, makeStyles, Theme, useTheme } from '@material-ui/core';
 import { ScreenTitle } from 'components/Common';
 import { DashboardTabPanel, DashboardUserNav, DashboardUserNavBtnType } from 'components/Dashboard';
+import { useSnackbar } from 'components/Feedback';
+import { useAuth } from 'core/api';
 import React, { FC, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { routes } from 'screens/consts';
@@ -16,15 +18,28 @@ type Props = StyleProps;
 
 export const DashboardEventsScreen: FC<Props> = () => {
   const [tab, setTab] = useState<number>(0);
+  const [sendEmailProcessing, setSendEmailProcessing] = useState<boolean>(false);
 
   const history = useHistory();
-
-  const isEmailConfirmed = false;
+  const { userConfirmed, resendEmailConfirmation } = useAuth();
+  const { showSnackbar } = useSnackbar();
 
   // Handlers
 
-  const handleResendEmailPress = () => {
+  const handleResendEmailPress = async () => {
     log.debug('hanlde resend email press');
+    try {
+      setSendEmailProcessing(true);
+      log.info('resending email confirmation');
+      await resendEmailConfirmation();
+      log.info('resending email confirmation done');
+      setSendEmailProcessing(false);
+      showSnackbar('Confirmation is sent. Please check your email.', 'success');
+    } catch (err) {
+      log.err(err);
+      setSendEmailProcessing(false);
+      showSnackbar(`Sending confirmation error`, 'error');
+    }
   };
 
   const handleUseNavBtnClick = (btn: DashboardUserNavBtnType) => {
@@ -55,7 +70,9 @@ export const DashboardEventsScreen: FC<Props> = () => {
         <Hidden smDown>{eventTabs}</Hidden>
         <Grid>
           <DashboardTabPanel className={classes.tabPanel} value={tab} index={0}>
-            {!isEmailConfirmed && <DashboardEmailConfirmScene onSubmit={handleResendEmailPress} />}
+            {!userConfirmed && (
+              <DashboardEmailConfirmScene processing={sendEmailProcessing} onSubmit={handleResendEmailPress} />
+            )}
             <Hidden smDown>
               <DashboardUserNav disabledBtns={['add']} onBtnClick={handleUseNavBtnClick} />
             </Hidden>